requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0
deep-translator>=1.11.4
Faker>=19.13.0
torch>=2.1
TTS>=0.21.4